Meeste account dingen zitten er nu in

+++ src/main/java/com/example/CodeGeneratieRestAPI/controllers/UserController.java
@@ -37,7 +37,7 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public User update(@PathVariable Integer id, @RequestBody User user) {
+    public User update(@PathVariable Long id, @RequestBody User user) {
         user.setId(id);
         return userService.update(user);
     }
